requirements.txt
matplotlib==3.10.8
pyserial==3.5
